+++ packages/components/scrollbar/src/Scrollbar.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 import Icon, { SpinnerM } from '@synerise/ds-icon';
+import theme from '@synerise/ds-core/dist/js/DSProvider/ThemeProvider/theme';
+
 import * as S from './Scrollbar.styles';
 import { ScrollbarProps } from './Scrollbar.types';
 import { DnDScrollbar } from './DnDScrollbar';
@@ -17,17 +19,17 @@ const Scrollbar = React.forwardRef<HTMLElement, ScrollbarProps>(
       );
     }, [children, fetchData, forwardedRef, props]);
 
-    return fetchData ? (
+    return (
       <S.ScrollbarContainer>
         {renderScrollbar}
         {loading && (
-          <S.Loader loading={loading}>
-            <Icon component={<SpinnerM />} color="#6a7580" />
-          </S.Loader>
+          <S.LoaderWrapper>
+            <S.Loader loading={loading}>
+              <Icon component={<SpinnerM />} color={theme.palette['grey-600']} />
+            </S.Loader>
+          </S.LoaderWrapper>
         )}
       </S.ScrollbarContainer>
-    ) : (
-      renderScrollbar
     );
   }
 );