+++ packages/portal/stories/components/ItemPicker/dataset.tsx
@@ -13,7 +13,11 @@ export const dataSource = [
   },
   {
     text: 'iPhone X',
-    prefixel: <Avatar src={imgSrc} size='small'>M</Avatar>,
+    prefixel: (
+      <Avatar src={imgSrc} size="small">
+        M
+      </Avatar>
+    ),
   },
   {
     text: 'MacBook Pro 15',
@@ -21,13 +25,44 @@ export const dataSource = [
   },
   {
     text: 'MacBook Air 13',
-    prefixel: <Avatar size="small" shape="square" backgroundColor='green'>E</Avatar>,
+    prefixel: (
+      <Avatar size="small" shape="square" backgroundColor="green">
+        E
+      </Avatar>
+    ),
   },
   {
     text: 'Macbook Pro 15',
     prefixel: <Icon component={<LaptopM />} color={theme.palette['grey-600']} />,
   },
   {
-    text: 'iPad Air 3'
-  }
+    text: 'iPad Air 3',
+  },
+  {
+    text: 'iPhone 13',
+    prefixel: <Icon component={<MobileM />} color={theme.palette['grey-600']} />,
+  },
+  {
+    text: 'iPhone 14',
+    prefixel: (
+      <Avatar src={imgSrc} size="small">
+        M
+      </Avatar>
+    ),
+  },
+  {
+    text: 'MacBook Pro 15 2023',
+    prefixel: <DSFlag country="US" />,
+  },
+  {
+    text: 'MacBook Air 13 2023',
+    prefixel: (
+      <Avatar size="small" shape="square" backgroundColor="green">
+        E
+      </Avatar>
+    ),
+  },
+  {
+    text: 'iPad Pro',
+  },
 ];