+++ packages/components/alert/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.28](https://github.com/Synerise/synerise-design/compare/@synerise/ds-alert@0.7.27...@synerise/ds-alert@0.7.28) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-alert
+
+
+
+
+
 ## [0.7.27](https://github.com/Synerise/synerise-design/compare/@synerise/ds-alert@0.7.26...@synerise/ds-alert@0.7.27) (2023-03-06)
 
 **Note:** Version bump only for package @synerise/ds-alert