+++ packages/components/avatar-group/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.72](https://github.com/Synerise/synerise-design/compare/@synerise/ds-avatar-group@0.6.71...@synerise/ds-avatar-group@0.6.72) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-avatar-group
+
+
+
+
+
+## [0.6.71](https://github.com/Synerise/synerise-design/compare/@synerise/ds-avatar-group@0.6.70...@synerise/ds-avatar-group@0.6.71) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-avatar-group
+
+
+
+
+
 ## [0.6.70](https://github.com/Synerise/synerise-design/compare/@synerise/ds-avatar-group@0.6.69...@synerise/ds-avatar-group@0.6.70) (2023-03-31)
 
 **Note:** Version bump only for package @synerise/ds-avatar-group