+++ packages/components/context-selector/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-context-selector@0.16.8...@synerise/ds-context-selector@0.16.9) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-context-selector
+
+
+
+
+
 ## [0.16.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-context-selector@0.16.7...@synerise/ds-context-selector@0.16.8) (2023-03-31)
 
 **Note:** Version bump only for package @synerise/ds-context-selector