+++ packages/components/date-range-picker/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.20.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-date-range-picker@0.20.8...@synerise/ds-date-range-picker@0.20.9) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-date-range-picker
+
+
+
+
+
+## [0.20.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-date-range-picker@0.20.7...@synerise/ds-date-range-picker@0.20.8) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-date-range-picker
+
+
+
+
+
 ## [0.20.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-date-range-picker@0.20.6...@synerise/ds-date-range-picker@0.20.7) (2023-03-31)