+++ packages/components/divider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.33](https://github.com/Synerise/synerise-design/compare/@synerise/ds-divider@0.5.32...@synerise/ds-divider@0.5.33) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-divider
+
+
+
+
+
 ## [0.5.32](https://github.com/Synerise/synerise-design/compare/@synerise/ds-divider@0.5.31...@synerise/ds-divider@0.5.32) (2023-03-31)
 
 **Note:** Version bump only for package @synerise/ds-divider