+++ packages/components/information-card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-information-card@0.3.8...@synerise/ds-information-card@0.3.9) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-information-card
+
+
+
+
+
 ## [0.3.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-information-card@0.3.7...@synerise/ds-information-card@0.3.8) (2023-03-31)
 
 **Note:** Version bump only for package @synerise/ds-information-card