+++ packages/components/input/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.18.37](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.18.36...@synerise/ds-input@0.18.37) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-input
+
+
+
+
+
 ## [0.18.36](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.18.35...@synerise/ds-input@0.18.36) (2023-03-06)
 
 **Note:** Version bump only for package @synerise/ds-input