+++ packages/components/search/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.46](https://github.com/Synerise/synerise-design/compare/@synerise/ds-search@0.8.45...@synerise/ds-search@0.8.46) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-search
+
+
+
+
+
 ## [0.8.45](https://github.com/Synerise/synerise-design/compare/@synerise/ds-search@0.8.44...@synerise/ds-search@0.8.45) (2023-03-31)
 
 **Note:** Version bump only for package @synerise/ds-search