+++ packages/components/select/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-select@0.15.1...@synerise/ds-select@0.15.2) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-select
+
+
+
+
+
 ## [0.15.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-select@0.15.0...@synerise/ds-select@0.15.1) (2023-03-09)