+++ packages/components/step-card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.17](https://github.com/Synerise/synerise-design/compare/@synerise/ds-step-card@0.8.16...@synerise/ds-step-card@0.8.17) (2023-04-05)
+
+**Note:** Version bump only for package @synerise/ds-step-card
+
+
+
+
+
 ## [0.8.16](https://github.com/Synerise/synerise-design/compare/@synerise/ds-step-card@0.8.15...@synerise/ds-step-card@0.8.16) (2023-03-31)
 
 **Note:** Version bump only for package @synerise/ds-step-card